+++ Intern/RayCore/src/Tracer/CpuTracer.cpp
@@ -26,9 +26,9 @@ std::vector<Ray> CpuTracer::trace(const Beamline& beamline) {
     auto rayList = beamline.getInputRays();
 
     CPU_TRACER::numberOfBeamlines = 1;
-    CPU_TRACER::numberOfElementsPerBeamline = beamline.m_OpticalElements.size();
-    CPU_TRACER::numberOfRays = rayList.size();
-    CPU_TRACER::numberOfRaysPerBeamLine = rayList.size();
+    CPU_TRACER::numberOfElementsPerBeamline = (double)beamline.m_OpticalElements.size();
+    CPU_TRACER::numberOfRays = (double)rayList.size();
+    CPU_TRACER::numberOfRaysPerBeamLine = (double)rayList.size();
 
     CPU_TRACER::quadricData.data.clear();
     CPU_TRACER::xyznull.data.clear();
